removing extra namespaces from services
