+++ islandora.module
@@ -6,7 +6,7 @@
  */
 
 use Drupal\islandora\DublinCore;
-use Drupal\islandora\Authentication\Provider\IslandoraTokenAuth;
+use Drupal\islandora\Authentication\Provider\TokenAuth;
 
 // @codingStandardsIgnoreStart
 // Common datastreams.
@@ -249,7 +249,7 @@ function islandora_user_access_check($object_or_datastream, array $permissions, 
 
   // Determine the user account to test against.
   if (!isset($user)) {
-    $token = filter_input(INPUT_GET, IslandoraTokenAuth::TOKEN_NAME, FILTER_SANITIZE_STRING);
+    $token = filter_input(INPUT_GET, TokenAuth::TOKEN_NAME, FILTER_SANITIZE_STRING);
     if ($token) {
       module_load_include('inc', 'islandora', 'includes/authtokens');
       $token_user = islandora_validate_object_token($object->id, $datastream->id, $token);
@@ -315,7 +315,7 @@ function islandora_object_datastream_tokened_access_callback($perm, $object = NU
   module_load_include('inc', 'islandora', 'includes/utilities');
 
   $token_account = NULL;
-  $token = filter_input(INPUT_GET, IslandoraTokenAuth::TOKEN_NAME, FILTER_SANITIZE_STRING);
+  $token = filter_input(INPUT_GET, TokenAuth::TOKEN_NAME, FILTER_SANITIZE_STRING);
 
   if ($token) {
     $user = islandora_validate_object_token($object->id, $datastream->id, $token);
@@ -504,7 +504,7 @@ function islandora_object_load($object_id) {
  */
 function islandora_tokened_object_load($object_id, array $map) {
   if (array_key_exists('token', $_GET)) {
-    $token = filter_input(INPUT_GET, IslandoraTokenAuth::TOKEN_NAME, FILTER_SANITIZE_STRING);
+    $token = filter_input(INPUT_GET, TokenAuth::TOKEN_NAME, FILTER_SANITIZE_STRING);
     if ($token) {
       module_load_include('inc', 'islandora', 'includes/authtokens');
       $token_user = islandora_validate_object_token($object_id, $map[4], $token);
@@ -743,7 +743,7 @@ function islandora_delete_datastream(AbstractDatastream &$datastream) {
  * Implements hook_cron().
  */
 function islandora_cron() {
-  IslandoraTokenAuth::removeExpiredTokens();
+  TokenAuth::removeExpiredTokens();
 }
 
 /**
