+++ islandora.services.yml
@@ -1,24 +1,24 @@
 services:
   object:
-    class: Drupal\islandora\ParamConverter\IslandoraObjectParamConverter
+    class: Drupal\islandora\ParamConverter\ObjectParamConverter
     arguments: ['@config.factory']
     tags:
       - { name: paramconverter }
   access_check.object:
-    class: Drupal\islandora\Access\IslandoraObjectAccess
+    class: Drupal\islandora\Access\ObjectAccess
     arguments: ['@config.factory']
     tags:
       - { name: access_check, applies_to: _islandora_object_access }
   datastream:
-    class: Drupal\islandora\ParamConverter\IslandoraDatastreamParamConverter
+    class: Drupal\islandora\ParamConverter\DatastreamParamConverter
     tags:
       - { name: paramconverter }
   access_check.datastream:
-    class: Drupal\islandora\Access\IslandoraDatastreamAccess
+    class: Drupal\islandora\Access\DatastreamAccess
     tags:
       - { name: access_check, applies_to: _islandora_datastream_access }
   authentication.islandora_token:
-    class: Drupal\islandora\Authentication\Provider\IslandoraTokenAuth
+    class: Drupal\islandora\Authentication\Provider\TokenAuth
     arguments: ['@config.factory', '@entity.manager']
     tags:
       - {name: authentication_provider, provider_id: islandora_auth_token, global: TRUE, priority: 1}
